storeAs support for dataReducer - #16
* Comments added.
